Implement std::error::Error for a few error enums
This has already been done for a few errors. It makes it so much
easier to work with `rust-dlc` in combination with a crate such as
`anyhow`.

Wherever we now have a source for an error variant, we remove the
source from the `Display` implementation to avoid redundancy. This
approach is included in the guidelines for implementing `Display::fmt`
and `Error::source`[1].

[1]:
https://blog.rust-lang.org/inside-rust/2021/07/01/What-the-error-handling-project-group-is-working-towards.html#guidelines-for-implementing-displayfmt-and-errorsource.