fixup! fix: Drop pending finalize messages on disconnected peer
The reconnect logic for a subchannel in the state of CloseConfirmed does not revert to an CloseOffered state, so that dropping the `CloseFinalize` message upon channel reestablishment turns out to be a regression. I haven't looked into the details, but I am wondering why the counter state is considered for the reconnect handling in the close protocol and not in the open protocol.