Fix fund output calculations
We were doing a couple of things wrong with the `extra_fee`, which is
normally used to account for potential buffer transaction fees when
setting up a DLC channel:

- We were not splitting the cost of the `extra_fee` between both
parties. Instead, we were accounting for it _twice_, deducted in full
from both the change output of the offer party and the accept party.

- We were using the "extra" `extra_fee` to pay more transaction fees
for the funding transaction. This is usually a completely unnecessary
overpayment.