Include base_weight_wu get_utxos_for_amount trait method
Before this patch, we were including the weight of the funding
transaction[^1] when calculating the `amount` that we pass to
`get_utxos_for_amount`.

This was fine, but the weight of the funding transaction directly
impacts the transaction fee that the consumer of `rust-dlc` will pay
as soon as the channel is opened on-chain. This means that, before
this patch, the bulk of the transaction fee was implicit in the amount
that the consumer was selecting coins for.

That behaviour is a bit surprising and it obfuscates the transaction
fee to be paid, which is important information to be shown to the end
user. Therefore, this patch adds a `base_weight_wu` which consumers
must use to calculate the funding transaction fee during coin
selection.

[^1]: It's not correct, but the intention was there.