Calculate expected DLC funding fees correctly
Inspired by this patch[^1] in `master`.

Since in the previous patch the `FUND_TX_BASE_WEIGHT` is passed
explicit to the implementer of `get_utxos_for_amount`, we can exclude
that from the target `amount` that we pass to `get_utxos_for_amount`.

Besides the party's collateral and all the fees related to opening a
_channel_, the only thing other that we need to take into account is
the transaction fee for the CET or the refund transaction.

The base weight can be split in half between the two parties, and each
party then has to pay for their own (possible) output on the CET or
refund transaction.

When constructing the fund transaction itself, we were also forgetting
to add the `8 + 1` bytes of each CET or refund transaction party
outputs.

[^1]: https://github.com/p2pderivatives/rust-dlc/pull/111/commits/c4a0f8820ea0531df9a7ce9e5705f1bdb098471d.