Do less in sub_channel_manager periodic check
Most users don't even need the `sub_channel_manager` periodic check to
run. But some still do, because there are pending LN-DLC channels
being closed in production.

So we still have to run this periodic check, but it's generally less
critical. It included a regular `channel_manager` periodic check to
make it more convenient for consumers to only have to run the
`sub_channel_manager` periodic check.

We now remove it because in 10101 we now want to execute the
`channel_manager` periodic check independently. It is therefore not
needed here. We remove it to avoid any unforeseen problems.